// src/app/interfaces/pokemon.Interfaces.ts
export interface DataToFront{
  next:string | null;
  prev:string | null;
  pokemons:Pokemon[]

}
export interface Pokemon{
  id:string;
  name:string;
  img:string;
}